scraper/requirements.txt
httpx>=0.25.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pyyaml>=6.0
requests>=2.31.0
fake-useragent>=1.4.0
tenacity>=8.2.0
python-dateutil>=2.8.0
